fix fps data struct reflection | fix system component, clean up, adjust to config

+++ Gems/FPSProfiler/Code/Source/Tools/FPSProfilerData.h
@@ -15,7 +15,7 @@ namespace FPSProfiler
         AZ::IO::Path m_OutputFilename = "@user@/fps_log.csv";
         bool m_SaveWithTimestamp = true;
         bool m_AutoSave = true;
-        float m_AutoSaveOccurrences = 100.0f;
+        int m_AutoSaveOccurrences = 100;
         float m_NearZeroPrecision = 0.01f;
         bool m_SaveFPSData = true;
         bool m_SaveGPUData = true;
